Use Fabric3 (Python 2 & 3 support)
---- Summary ----

Updated requirements and `fab serve()` implementation. Fixed all Python
2 & 3 inconsistentes in the output. As a side effect the description
escaping and the canonical link fixes was created :)

---- Details ----

Fabric3 is transparent replacement for Fabric, which supporting Python 2
& 3 both. The transition was very smooth for Fabric itself: just update
requirements.txt and fix `fab serve()` to run python2 or python3 in
dependence of current interpteter.

During testing I found the inconsistence in htmlmin across Python major
versions. It replaces non-breaking and thin whitespaces, that utf-8
encoded (not html entities), with the regular whitespaces when Python 3
used. Reported it in [1]. Anyway, I replaced all unicode whitespaces in
our articles using our backslash-whitespace syntax.

Found that 'Description' meta value which contains double quotes
produces malformed 'og:description' meta tag. Enable escaping for
'Description' and 'Title' everywhere it used as an attribute value.
Anyway, I replaced all double quotes in our descriptions with typography
equivalents.

Why `str1|safe + str2|e + str3|safe` is used? The `safe` keywords in
right and left strings is necessary to prevent its escaping due to the
way Jinja handle escaping. Espaced string is not just string, it'is
Markup object--that's for prevent double escaping. But contatenation
(`+`) regular string to Markup object causes escaping the string. That's
why the `safe` is necessary.

The another one inconsistence was found and looks like t's from the
Pelican side. When the `category` value accessed from Jinja template
under Python 2 it was 'What If?', not 'what-if', and the category check
was failing. The `category.slug` works good in both Python versions, so
now the canonical path for the `/what-if/` page is the main page and not
`/what-if/` itself despite which Python used.

[1]: https://github.com/mankyd/htmlmin/issues/39