Prevent horizontal scrollbar from appear
It was when a foonote opened in the small layout and the screen (the
viewport) is small enough.

It looks like this is the side effect of using negative `left`
attribute: the scrollbar appears exactly when a element wouldn't fit the
viewport with `left` set to 0. To more on that: new empty space on the
right side of viewport exactly corresponds the right side of the element
with `left` set to 0.