Add options to handle Pelican's RSS changes
It gives back full content in RSS feeds. This was changed to
summary-only in [1]. The option was added in [2]. Both between 3.6.3 and
3.7.

The only changes in feeds between 3.6.3 (with hardcoded html5 as
markdown's output format) and 3.7 are the following:

* Atom feeds: full content of article moved from <summary/> to
  </content> and <summary/> becomes really summary.
* Atom/RSS feeds: strings in <id/> / <guid/> slightly changed by [3].

The first should be good, the second can cause some RSS/Atom readers to
find fake new items when the website will updated.

[1]: https://github.com/getpelican/pelican/pull/1989
[2]: https://github.com/getpelican/pelican/pull/2051
[3]: https://github.com/getpelican/pelican/pull/1306